+++ apps/workflows/tests/test_models.py
@@ -911,7 +911,7 @@ class TestFlaw:
                 "description": "random description",
                 "priority": 1,  # is more prior than default one
                 "conditions": [
-                    "major_incident_state_is_approved"
+                    "major_incident_state_is_major_incident_approved"
                 ],  # major incident flaws are classified here
                 "states": [],  # this is not valid but OK for this test
             }
@@ -919,13 +919,15 @@ class TestFlaw:
         workflow.states = states
         workflow_framework.register_workflow(workflow)
 
-        flaw = FlawFactory(major_incident_state=Flaw.FlawMajorIncident.APPROVED)
+        flaw = FlawFactory(
+            major_incident_state=Flaw.FlawMajorIncident.MAJOR_INCIDENT_APPROVED
+        )
         AffectFactory(flaw=flaw)
         flaw.adjust_classification()
 
         assert flaw.classification["workflow"] == "MAJOR_INCIDENT"
 
-        flaw.major_incident_state = Flaw.FlawMajorIncident.REJECTED
+        flaw.major_incident_state = Flaw.FlawMajorIncident.MAJOR_INCIDENT_REJECTED
         flaw.adjust_classification()
 
         assert flaw.classification["workflow"] == "DEFAULT"